pyproject.toml
[tool.poetry]
name = "thebaeoom-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.6"
djangorestframework = "^3.13.1"
mysqlclient = "^2.1.1"
django-environ = "^0.9.0"
django-cors-headers = "^3.13.0"
django-admin-json-editor = "^0.2.3"
django-filter = "^22.1"
django-ckeditor = "^6.4.2"
Pillow = "^9.2.0"
django-admin-interface = "^0.19.1"
django-jsonform = "^2.10.1"
django-ajax-selects = "^2.2.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
